seed.db only removes admin teachers before seeding

+++ db/seeds.rb
@@ -1,6 +1,6 @@
 require_relative 'seed_data'
 
-Teacher.destroy_all
+Teacher.where(admin: true).destroy_all
 EmailTemplate.destroy_all
 
 SeedData.emails.each {|email| EmailTemplate.find_or_create_by(email)}
